+++ resources/libraries/rbmonitor_helpers.rb
@@ -102,6 +102,7 @@ module Rbmonitor
 
       monitors = []
       inserted = {}
+      inserted_operations = []
       send_flag = 0
 
       begin
@@ -113,11 +114,12 @@ module Rbmonitor
       resource_node['redborder']['monitors'].each do |resource_node_monitor|
         monitor = resource_node_monitor.to_hash
         name    = monitor['name']
+        operation = monitor['system']
         next unless name
 
-        # skip if already inserted or not allowed by data_bag
+        # skip if already inserted with the same operation if has operation or not allowed by data_bag
         monitors_list = data_bag['monitors']
-        next unless inserted[name].nil? && (monitors_list.nil? || monitors_list.include?(name))
+        next unless (inserted[name].nil? || !inserted_operations.include?(operation)) && (monitors_list.nil? || monitors_list.include?(name))
 
         # decide send_flag for this monitor
         send_flag = resource_node['redborder']['monitors'].any? do |m|
@@ -179,11 +181,17 @@ module Rbmonitor
             val.gsub!('rb_get_redfish.sh', cmd)
           end
 
+          # Format monitor enrichment as a correct JSON being a Ruby hash if is a endpoint
+          if monitor[k].is_a?(Hash) && !monitor[k]['endpoint'].nil?
+            val = monitor[k]
+          end
+
           monitor[k] = val
         end
 
         monitor['send'] = send_flag
         inserted[name]  = true
+        inserted_operations << operation
         monitors << monitor
       end